Refuse to generate a patch for a branch that is not checked out.
Normally, one wants only to generate the patches for against the base
revision of the currently checked out branch. We therefore refuse
doing anything differently from now on.

This makes the branch argument redundant but later we want to allow that
case via a force option or something like this.